Update bridge generator

+++ Code/Server/core/entities.h
@@ -107,7 +107,7 @@ oak_object *oak_entity_get(oak_type type, u32 id) {
 int oak_entity_get_id_from_librg(librg_entity *entity) {
     ZPL_ASSERT_NOT_NULL(entity);
 
-    return (int)(entity->user_data);
+    return (int)(uintptr)(entity->user_data);
 }
 
 /**
